Rename fault-count a mod1 folder and file
