+++ src/providers/fault-count/cat-a-mod1/fault-count.cat-a-mod1.ts
