+++ src/providers/fault-count/fault-count.ts
@@ -9,7 +9,7 @@ import { FaultCountBEHelper } from './cat-be/fault-count.cat-be';
 import { FaultCountCHelper } from './cat-c/fault-count.cat-c';
 
 import { sumManoeuvreFaults } from '../../shared/helpers/faults';
-import { FaultCountAM1Helper } from './cat-am1/fault-count.cat-a';
+import { FaultCountAM1Helper } from './cat-a-mod1/fault-count.cat-a-mod1';
 
 @Injectable()
 export class FaultCountProvider {
